requirements.txt
dnspython==2.7.0
Flask==3.1.0
flask_minify==0.48
flask_restful==0.3.10